+++ tests/extra/job_management/test_stac_job_db.py
@@ -1,6 +1,7 @@
+import collections
 import datetime
 import re
-from typing import Any, Dict
+from typing import Any, Dict, List, Optional, Union
 from unittest import mock
 from unittest.mock import MagicMock, patch
 
@@ -11,7 +12,6 @@ import pandas.testing as pdt
 import pystac
 import pystac_client
 import pytest
-from requests.auth import AuthBase
 from shapely.geometry import Point
 
 from openeo.extra.job_management import MultiBackendJobManager
@@ -20,27 +20,13 @@ from openeo.rest._testing import DummyBackend
 
 
 @pytest.fixture
-def mock_auth():
-    return MagicMock(spec=AuthBase)
-
-
-@pytest.fixture
-def mock_stac_api_job_database(mock_auth) -> STACAPIJobDatabase:
-    return STACAPIJobDatabase(collection_id="test_id", stac_root_url="http://fake-stac-api.test", auth=mock_auth)
-
-
-@pytest.fixture
-def mock_pystac_client(dummy_stac_item):
+def mock_pystac_client():
     mock_client = MagicMock(spec=pystac_client.Client)
 
     mock_client.get_collections.return_value = [
         MagicMock(id="collection-1"),
         MagicMock(id="collection-2"),
     ]
-
-    mock_item_search = MagicMock(spec=pystac_client.ItemSearch)
-    mock_item_search.items.return_value = [dummy_stac_item]
-    mock_client.search.return_value = mock_item_search
 
     with patch("pystac_client.Client.open", return_value=mock_client):
         yield mock_client
@@ -65,164 +51,56 @@ def job_db_not_exists(mock_pystac_client) -> STACAPIJobDatabase:
     )
 
 
-@pytest.fixture
-def dummy_dataframe() -> pd.DataFrame:
-    return pd.DataFrame({"no": [1], "geometry": [2], "here": [3]})
-
-
-@pytest.fixture
-def normalized_dummy_dataframe() -> pd.DataFrame:
-    return pd.DataFrame(
-        {
-            "item_id": [0],
-            "no": [1],
-            "geometry": [2],
-            "here": [3],
-            "id": None,
-            "backend_name": None,
-            "status": ["not_started"],
-            "start_time": None,
-            "running_start_time": None,
-            "cpu": None,
-            "memory": None,
-            "duration": None,
-            "costs": None,
-        },
-    )
-
-
-@pytest.fixture
-def another_dummy_dataframe() -> pd.DataFrame:
-    return pd.DataFrame({"item_id": [1], "no": [4], "geometry": [5], "here": [6]})
-
-
-@pytest.fixture
-def normalized_merged_dummy_dataframe() -> pd.DataFrame:
-    return pd.DataFrame(
-        {
-            "item_id": [0, 1],
-            "no": [1, 4],
-            "geometry": [2, 5],
-            "here": [3, 6],
-            "id": None,
-            "backend_name": None,
-            "status": ["not_started", "not_started"],
-            "start_time": None,
-            "running_start_time": None,
-            "cpu": None,
-            "memory": None,
-            "duration": None,
-            "costs": None,
-        }
-    )
-
-
-@pytest.fixture
-def dummy_geodataframe() -> gpd.GeoDataFrame:
-    return gpd.GeoDataFrame(
-        {
-            "there": [1],
-            "is": [2],
-            "geometry": [Point(1, 1)],
-        },
-        geometry="geometry",
-    )
-
-
-@pytest.fixture
-def normalized_dummy_geodataframe() -> pd.DataFrame:
-    return pd.DataFrame(
-        {
-            "item_id": [0],
-            "there": [1],
-            "is": [2],
-            "geometry": [{"type": "Point", "coordinates": (1.0, 1.0)}],
-            "id": None,
-            "backend_name": None,
-            "status": ["not_started"],
-            "start_time": None,
-            "running_start_time": None,
-            "cpu": None,
-            "memory": None,
-            "duration": None,
-            "costs": None,
-        }
-    )
-
-
-FAKE_NOW = datetime.datetime(2020, 5, 22)
-
-
-@pytest.fixture
-def dummy_stac_item() -> pystac.Item:
-    properties = {
-        "datetime": pystac.utils.datetime_to_str(FAKE_NOW),
-        "some_property": "value",
+def _common_normalized_df_data(rows: int = 1) -> dict:
+    """
+    Helper to build a dict (to be passed to `pd.DataFrame`)
+    with common columns that are the result of normalization.
+    In the context of these tests however, they are
+    mainly irrelevant boilerplate data that is tedious to repeat.
+    """
+    return {
+        "id": None,
+        "backend_name": None,
+        "status": ["not_started"] * rows,
+        "start_time": None,
+        "running_start_time": None,
+        "cpu": None,
+        "memory": None,
+        "duration": None,
+        "costs": None,
     }
 
-    return pystac.Item(id="test", geometry=None, bbox=None, properties=properties, datetime=FAKE_NOW)
 
+def _pystac_item(
+    *,
+    id: str,
+    properties: Optional[dict] = None,
+    geometry: Optional = None,
+    bbox: Optional = None,
+    datetime_: Union[None, str, datetime.date, datetime.date] = "2025-06-07",
+    links: Optional[List[Union[pystac.Link, dict]]] = None,
+    **kwargs,
+) -> pystac.Item:
+    """Helper to easily construct a dummy but valid pystac.Item"""
+    if isinstance(datetime_, str):
+        datetime_ = pystac.utils.str_to_datetime(datetime_)
+    elif isinstance(datetime_, datetime.date):
+        datetime_ = datetime.datetime.combine(datetime_, datetime.time.min, tzinfo=datetime.timezone.utc)
 
-@pytest.fixture
-def dummy_stac_item_geometry() -> pystac.Item:
-    properties = {
-        "datetime": pystac.utils.datetime_to_str(FAKE_NOW),
-        "some_property": "value",
-        "geometry": {"type": "Polygon", "coordinates": (((0.0, 0.0), (1.0, 0.0), (1.0, 1.0), (0.0, 1.0), (0.0, 0.0)),)},
-    }
-
-    return pystac.Item(
-        id="test",
-        geometry={"type": "Polygon", "coordinates": (((0.0, 0.0), (1.0, 0.0), (1.0, 1.0), (0.0, 1.0), (0.0, 0.0)),)},
-        bbox=(0.0, 0.0, 1.0, 1.0),
-        properties=properties,
-        datetime=FAKE_NOW,
+    item = pystac.Item(
+        id=id,
+        geometry=geometry,
+        bbox=bbox,
+        properties=properties or {},
+        datetime=datetime_,
+        **kwargs,
     )
 
+    if links:
+        for link in links:
+            item.add_link(pystac.Link.from_dict(link) if isinstance(link, dict) else link)
 
-@pytest.fixture
-def dummy_series() -> pd.Series:
-    return pd.Series(
-        {"item_id": "test", "datetime": pystac.utils.datetime_to_str(FAKE_NOW), "some_property": "value"}, name="test"
-    )
-
-
-@pytest.fixture
-def dummy_series_no_item_id() -> pd.Series:
-    return pd.Series({"datetime": pystac.utils.datetime_to_str(FAKE_NOW), "some_property": "value"}, name="test")
-
-@pytest.fixture
-def dummy_series_geometry() -> pd.Series:
-    return pd.Series(
-        {
-            "item_id": "test",
-            "datetime": pystac.utils.datetime_to_str(FAKE_NOW),
-            "some_property": "value",
-            "geometry": {
-                "type": "Polygon",
-                "coordinates": (((0.0, 0.0), (1.0, 0.0), (1.0, 1.0), (0.0, 1.0), (0.0, 0.0)),),
-            },
-        },
-        name="test",
-    )
-
-
-@pytest.fixture
-def patch_datetime_now():
-    with patch("datetime.datetime") as mock_datetime:
-        mock_datetime.now.return_value = FAKE_NOW
-        yield mock_datetime
-
-
-@pytest.fixture
-def bulk_dataframe():
-    return pd.DataFrame(
-        {
-            "item_id": [f"test-{i}" for i in range(10)],
-            "some_property": [f"value-{i}" for i in range(10)],
-        },
-        index=[i for i in range(10)],
-    )
+    return item
 
 
 class TestSTACAPIJobDatabase:
@@ -231,20 +109,45 @@ class TestSTACAPIJobDatabase:
         assert job_db_exists.exists() == True
         assert job_db_not_exists.exists() == False
 
+    @pytest.mark.parametrize(
+        ["df", "expected"],
+        [
+            (
+                pd.DataFrame({"no": [1], "geometry": [2], "here": [3]}),
+                pd.DataFrame(
+                    data={
+                        "item_id": [0],
+                        "no": [1],
+                        "geometry": [2],
+                        "here": [3],
+                        **_common_normalized_df_data(),
+                    },
+                ),
+            ),
+            (
+                pd.DataFrame({"item_id": ["item-123", "item-456"], "hello": ["world", "earth"]}),
+                pd.DataFrame(
+                    data={
+                        "item_id": ["item-123", "item-456"],
+                        "hello": ["world", "earth"],
+                        **_common_normalized_df_data(rows=2),
+                    },
+                ),
+            ),
+        ],
+    )
     @patch("openeo.extra.job_management.stac_job_db.STACAPIJobDatabase.persist", return_value=None)
-    def test_initialize_from_df_non_existing(
-        self, mock_persist, job_db_not_exists, dummy_dataframe, normalized_dummy_dataframe
-    ):
-
-        job_db_not_exists.initialize_from_df(dummy_dataframe)
+    def test_initialize_from_df_non_existing(self, mock_persist, job_db_not_exists, df, expected):
+        job_db_not_exists.initialize_from_df(df)
 
         mock_persist.assert_called_once()
-        pdt.assert_frame_equal(mock_persist.call_args[0][0], normalized_dummy_dataframe)
+        pdt.assert_frame_equal(mock_persist.call_args[0][0], expected)
         assert job_db_not_exists.has_geometry == False
 
-    def test_initialize_from_df_existing_error(self, job_db_exists, dummy_dataframe):
+    def test_initialize_from_df_existing_error(self, job_db_exists):
+        df = pd.DataFrame({"hello": ["world"]})
         with pytest.raises(FileExistsError):
-            job_db_exists.initialize_from_df(dummy_dataframe)
+            job_db_exists.initialize_from_df(df)
 
     @patch("openeo.extra.job_management.stac_job_db.STACAPIJobDatabase.persist", return_value=None)
     @patch("openeo.extra.job_management.stac_job_db.STACAPIJobDatabase.get_by_status")
@@ -253,45 +156,192 @@ class TestSTACAPIJobDatabase:
         mock_get_by_status,
         mock_persist,
         job_db_exists,
-        normalized_dummy_dataframe,
-        another_dummy_dataframe,
-        normalized_merged_dummy_dataframe,
     ):
-        mock_get_by_status.return_value = normalized_dummy_dataframe
-        job_db_exists.initialize_from_df(another_dummy_dataframe, on_exists="append")
+        mock_get_by_status.return_value = pd.DataFrame(
+            {
+                "item_id": [0],
+                "no": [1],
+                "geometry": [2],
+                "here": [3],
+                **_common_normalized_df_data(),
+            },
+        )
+
+        df = pd.DataFrame(
+            {
+                "item_id": [1],
+                "no": [4],
+                "geometry": [5],
+                "here": [6],
+            }
+        )
+        job_db_exists.initialize_from_df(df, on_exists="append")
 
         mock_persist.assert_called_once()
-        pdt.assert_frame_equal(mock_persist.call_args[0][0], normalized_merged_dummy_dataframe)
+        expected_df = pd.DataFrame(
+            {
+                "item_id": [0, 1],
+                "no": [1, 4],
+                "geometry": [2, 5],
+                "here": [3, 6],
+                **_common_normalized_df_data(rows=2),
+            }
+        )
+        pdt.assert_frame_equal(mock_persist.call_args[0][0], expected_df)
         assert job_db_exists.has_geometry == False
 
+    @pytest.mark.parametrize(
+        ["df", "expected"],
+        [
+            (
+                gpd.GeoDataFrame(
+                    {
+                        "there": [1],
+                        "is": [2],
+                        "geometry": [Point(1, 1)],
+                    },
+                    geometry="geometry",
+                ),
+                pd.DataFrame(
+                    {
+                        "item_id": [0],
+                        "there": [1],
+                        "is": [2],
+                        "geometry": [{"type": "Point", "coordinates": (1.0, 1.0)}],
+                        **_common_normalized_df_data(),
+                    }
+                ),
+            ),
+            (
+                gpd.GeoDataFrame(
+                    data={"item_id": ["item-123", "item-456"], "hello": ["world", "earth"]},
+                    geometry=[Point(1, 1), Point(2, 2)],
+                ),
+                pd.DataFrame(
+                    data={
+                        "item_id": ["item-123", "item-456"],
+                        "hello": ["world", "earth"],
+                        "geometry": [
+                            {"type": "Point", "coordinates": (1.0, 1.0)},
+                            {"type": "Point", "coordinates": (2.0, 2.0)},
+                        ],
+                        **_common_normalized_df_data(rows=2),
+                    },
+                ),
+            ),
+        ],
+    )
     @patch("openeo.extra.job_management.stac_job_db.STACAPIJobDatabase.persist", return_value=None)
-    def test_initialize_from_df_with_geometry(
-        self, mock_persists, job_db_not_exists, dummy_geodataframe, normalized_dummy_geodataframe
-    ):
-        job_db_not_exists.initialize_from_df(dummy_geodataframe)
+    def test_initialize_from_df_with_geometry(self, mock_persists, job_db_not_exists, df, expected):
+        job_db_not_exists.initialize_from_df(df)
 
         mock_persists.assert_called_once()
-        pdt.assert_frame_equal(mock_persists.call_args[0][0], normalized_dummy_geodataframe)
+        pdt.assert_frame_equal(mock_persists.call_args[0][0], expected)
         assert job_db_not_exists.has_geometry == True
         assert job_db_not_exists.geometry_column == "geometry"
 
-    def test_series_from(self, job_db_exists, dummy_series_no_item_id, dummy_stac_item):
-        pdt.assert_series_equal(job_db_exists.series_from(dummy_stac_item), dummy_series_no_item_id)
+    @pytest.mark.parametrize(
+        ["item", "expected"],
+        [
+            (
+                _pystac_item(
+                    id="item-123",
+                    properties={"some_property": "value"},
+                    datetime_=datetime.datetime(2020, 5, 22),
+                ),
+                pd.Series(
+                    name="item-123",
+                    data={"some_property": "value", "datetime": "2020-05-22T00:00:00Z"},
+                ),
+            ),
+        ],
+    )
+    def test_series_from(self, job_db_exists, item, expected):
+        pdt.assert_series_equal(job_db_exists.series_from(item), expected)
 
-    def test_item_from(self, patch_datetime_now, job_db_exists, dummy_series, dummy_stac_item):
-        item = job_db_exists.item_from(dummy_series)
-        assert item.to_dict() == dummy_stac_item.to_dict()
+    @pytest.mark.parametrize(
+        ["series", "expected"],
+        [
+            # Minimal (using current time as datetime)
+            (
+                pd.Series({"item_id": "item-123"}),
+                _pystac_item(id="item-123", datetime_="2022-02-02"),
+            ),
+            # With explicit datetime, and some other properties
+            (
+                pd.Series(
+                    {
+                        "item_id": "item-123",
+                        "datetime": "2023-04-05",
+                        "hello": "world",
+                    }
+                ),
+                _pystac_item(
+                    id="item-123",
+                    properties={"hello": "world"},
+                    datetime_="2023-04-05",
+                ),
+            ),
+            (
+                pd.Series(
+                    {
+                        "item_id": "item-123",
+                        "datetime": datetime.datetime(2023, 4, 5, 12, 34),
+                        "hello": "world",
+                    }
+                ),
+                _pystac_item(
+                    id="item-123",
+                    properties={"hello": "world"},
+                    datetime_="2023-04-05T12:34:00Z",
+                ),
+            ),
+        ],
+    )
+    def test_item_from(self, job_db_exists, series, expected, time_machine):
+        time_machine.move_to("2022-02-02T00:00:00Z")
+        item = job_db_exists.item_from(series)
+        assert isinstance(item, pystac.Item)
+        assert item.to_dict() == expected.to_dict()
 
-    def test_item_from_geometry(
-        self, patch_datetime_now, job_db_exists, dummy_series_geometry, dummy_stac_item_geometry
-    ):
+    @pytest.mark.parametrize(
+        ["series", "expected"],
+        [
+            (
+                pd.Series(
+                    {
+                        "item_id": "item-123",
+                        "datetime": "2023-04-05",
+                        "geometry": {"type": "Polygon", "coordinates": (((1, 2), (3, 4), (0, 5), (1, 2)),)},
+                    },
+                ),
+                _pystac_item(
+                    id="item-123",
+                    datetime_="2023-04-05",
+                    geometry={"type": "Polygon", "coordinates": (((1, 2), (3, 4), (0, 5), (1, 2)),)},
+                    bbox=(0.0, 2.0, 3.0, 5.0),
+                    properties={
+                        "geometry": {"type": "Polygon", "coordinates": (((1, 2), (3, 4), (0, 5), (1, 2)),)},
+                    },
+                ),
+            ),
+        ],
+    )
+    def test_item_from_geometry(self, job_db_exists, series, expected):
+        # TODO avoid this `has_geometry` hack?
         job_db_exists.has_geometry = True
-        item = job_db_exists.item_from(dummy_series_geometry)
-        assert item.to_dict() == dummy_stac_item_geometry.to_dict()
+        item = job_db_exists.item_from(series)
+        assert isinstance(item, pystac.Item)
+        assert item.to_dict() == expected.to_dict()
 
     @patch("openeo.extra.job_management.stac_job_db.STACAPIJobDatabase.get_by_status")
-    def test_count_by_status(self, mock_get_by_status, normalized_dummy_dataframe, job_db_exists):
-        mock_get_by_status.return_value = normalized_dummy_dataframe
+    def test_count_by_status(self, mock_get_by_status, job_db_exists):
+        mock_get_by_status.return_value = pd.DataFrame(
+            {
+                "item_id": [0],
+                "status": ["not_started"],
+            },
+        )
         assert job_db_exists.count_by_status() == {"not_started": 1}
 
     def test_get_by_status_no_filter(self, job_db_exists):
@@ -307,103 +357,129 @@ class TestSTACAPIJobDatabase:
             method="GET", collections=["collection-1"], filter="\"properties.status\"='not_started'", max_items=None
         )
 
-    def test_get_by_status_result(self, job_db_exists):
-        df = job_db_exists.get_by_status(["not_started"])
+    def test_get_by_status_result(self, requests_mock):
+        stac_api_url = "http://stacapi.test"
+        dummy_stac_api = DummyStacApi(root_url=stac_api_url, requests_mock=requests_mock)
+        dummy_stac_api.predefine_collection("collection-123")
+        dummy_stac_api.predefine_item(
+            collection_id="collection-123",
+            item=_pystac_item(id="item-123", properties={"status": "not_started"}),
+        )
+        dummy_stac_api.predefine_item(
+            collection_id="collection-123",
+            item=_pystac_item(id="item-456", properties={"status": "running"}),
+        )
+        dummy_stac_api.predefine_item(
+            collection_id="collection-123",
+            item=_pystac_item(id="item-789", properties={"status": "not_started"}),
+        )
+
+        job_db = STACAPIJobDatabase(collection_id="collection-123", stac_root_url=stac_api_url)
 
         pdt.assert_frame_equal(
-            df,
+            job_db.get_by_status(["not_started"]),
             pd.DataFrame(
                 {
-                    "item_id": ["test"],
-                    "datetime": [pystac.utils.datetime_to_str(FAKE_NOW)],
-                    "some_property": ["value"],
+                    "item_id": ["item-123", "item-789"],
+                    "status": ["not_started", "not_started"],
+                    "datetime": ["2025-06-07T00:00:00Z", "2025-06-07T00:00:00Z"],
                 },
-                index=[0],
+            ),
+        )
+        pdt.assert_frame_equal(
+            job_db.get_by_status(["running"]),
+            pd.DataFrame(
+                {
+                    "item_id": ["item-456"],
+                    "status": ["running"],
+                    "datetime": ["2025-06-07T00:00:00Z"],
+                },
             ),
         )
 
-    @patch("requests.post")
-    def test_persist_single_chunk(self, mock_requests_post, bulk_dataframe, job_db_exists, patch_datetime_now):
-        def bulk_items(df):
-            all_items = []
-            if not df.empty:
-
-                def handle_row(series):
-                    item = job_db_exists.item_from(series)
-                    job_db_exists._prepare_item(item, job_db_exists.collection_id)
-                    all_items.append(item)
-
-                df.apply(handle_row, axis=1)
-            return all_items
-
-        items = bulk_items(bulk_dataframe)
-
-        mock_requests_post.return_value.status_code = 200
-        mock_requests_post.return_value.json.return_value = {"status": "success"}
-        mock_requests_post.reason = "OK"
-
-        job_db_exists.persist(bulk_dataframe)
-
-        mock_requests_post.assert_called_once()
-
-        mock_requests_post.assert_called_with(
-            url=f"http://fake-stac-api/collections/{job_db_exists.collection_id}/bulk_items",
-            auth=None,
-            json={
-                "method": "upsert",
-                "items": {item.id: item.to_dict() for item in items},
+    def test_persist_single_chunk(self, requests_mock, job_db_exists):
+        rows = 5
+        bulk_dataframe = pd.DataFrame(
+            data={
+                "item_id": [f"item-{i}" for i in range(rows)],
+                "datetime": [f"2020-{i + 1:02d}-01" for i in range(rows)],
+                "some_property": [f"value-{i}" for i in range(rows)],
             },
         )
 
-    @patch("requests.post")
-    def test_persist_multiple_chunks(self, mock_requests_post, bulk_dataframe, job_db_exists):
-        def bulk_items(df):
-            all_items = []
-            if not df.empty:
-
-                def handle_row(series):
-                    item = job_db_exists.item_from(series)
-                    job_db_exists._prepare_item(item, job_db_exists.collection_id)
-                    all_items.append(item)
-
-                df.apply(handle_row, axis=1)
-            return all_items
-
-        items = bulk_items(bulk_dataframe)
-
-        mock_requests_post.return_value.status_code = 200
-        mock_requests_post.return_value.json.return_value = {"status": "success"}
-        mock_requests_post.reason = "OK"
-
-        job_db_exists.bulk_size = 3
-        job_db_exists._upload_items_bulk(collection_id=job_db_exists.collection_id, items=items)
-
-        # 10 items in total, 3 items per chunk, should result in 4 calls
-        assert sorted(
-            (c.kwargs for c in mock_requests_post.call_args_list),
-            key=lambda d: sorted(d["json"]["items"].keys()),
-        ) == [
-            {
-                "url": f"http://fake-stac-api/collections/{job_db_exists.collection_id}/bulk_items",
-                "auth": None,
-                "json": {"method": "upsert", "items": {item.id: item.to_dict() for item in items[:3]}},
-            },
-            {
-                "url": f"http://fake-stac-api/collections/{job_db_exists.collection_id}/bulk_items",
-                "auth": None,
-                "json": {"method": "upsert", "items": {item.id: item.to_dict() for item in items[3:6]}},
-            },
-            {
-                "url": f"http://fake-stac-api/collections/{job_db_exists.collection_id}/bulk_items",
-                "auth": None,
-                "json": {"method": "upsert", "items": {item.id: item.to_dict() for item in items[6:9]}},
-            },
-            {
-                "url": f"http://fake-stac-api/collections/{job_db_exists.collection_id}/bulk_items",
-                "auth": None,
-                "json": {"method": "upsert", "items": {item.id: item.to_dict() for item in items[9:]}},
-            },
+        expected_items = [
+            _pystac_item(
+                id=f"item-{i}",
+                properties={"some_property": f"value-{i}"},
+                datetime_=f"2020-{i + 1:02d}-01",
+                collection="collection-1",
+                links=[{"rel": "collection", "href": "collection-1"}],
+            )
+            for i in range(rows)
         ]
+        expected_items = {item.id: item.to_dict() for item in expected_items}
+
+        def post_bulk_items(request, context):
+            post_data = request.json()
+            assert post_data == {"method": "upsert", "items": expected_items}
+            return {"status": "success"}
+
+        post_bulk_items_mock = requests_mock.post(
+            re.compile(r"http://fake-stac-api/collections/.*/bulk_items"), json=post_bulk_items
+        )
+
+        job_db_exists.persist(bulk_dataframe)
+        assert post_bulk_items_mock.called
+
+
+
+    def test_persist_multiple_chunks(self, requests_mock, job_db_exists):
+        rows = 12
+        bulk_dataframe = pd.DataFrame(
+            data={
+                "item_id": [f"item-{i}" for i in range(rows)],
+                "datetime": [f"2020-{i + 1:02d}-01" for i in range(rows)],
+                "some_property": [f"value-{i}" for i in range(rows)],
+            },
+        )
+
+        chunks = []
+
+        def post_bulk_items(request, context):
+            nonlocal chunks
+            post_data = request.json()
+            chunks.append(post_data)
+            return {"status": "success"}
+
+        post_bulk_items_mock = requests_mock.post(
+            re.compile(r"http://fake-stac-api/collections/.*/bulk_items"), json=post_bulk_items
+        )
+
+        job_db_exists.bulk_size = 5
+        job_db_exists.persist(bulk_dataframe)
+
+        assert post_bulk_items_mock.call_count == 3
+        expected = [
+            {
+                "method": "upsert",
+                "items": {
+                    f"item-{i}": _pystac_item(
+                        id=f"item-{i}",
+                        properties={"some_property": f"value-{i}"},
+                        datetime_=f"2020-{i + 1:02d}-01",
+                        collection="collection-1",
+                        links=[{"rel": "collection", "href": "collection-1"}],
+                    ).to_dict()
+                    for i in range(start, end)
+                },
+            }
+            for (start, end) in [(0, 5), (5, 10), (10, 12)]
+        ]
+
+        def normalize(data):
+            return sorted(data, key=lambda d: min(d["items"].keys()))
+
+        assert normalize(chunks) == normalize(expected)
 
 
 @pytest.fixture
@@ -422,16 +498,16 @@ def sleep_mock():
 class DummyStacApi:
     """Minimal dummy implementation of a STAC API for testing purposes."""
 
-    def __init__(self, root_url: str, requests_mock):
+    def __init__(self, *, root_url: str = "http://stacapi.test", requests_mock):
         self.root_url = root_url.rstrip("/")
         self._requests_mock = requests_mock
 
         requests_mock.get(f"{self.root_url}/", json=self._get_root())
-        self.collections = []
+        self.collections: List[dict] = []
         requests_mock.get(f"{self.root_url}/collections", json=self._get_collections)
         requests_mock.post(f"{self.root_url}/collections", json=self._post_collections)
 
-        self.items: Dict[str, Dict[str, Any]] = {}
+        self.items: Dict[str, Dict[str, Any]] = collections.defaultdict(dict)
         requests_mock.post(
             re.compile(rf"{self.root_url}/collections/[^/]+/bulk_items"), json=self._post_collections_bulk_items
         )
@@ -464,6 +540,26 @@ class DummyStacApi:
         self.collections.append(post_data)
         return {}
 
+    def predefine_collection(self, collection_id: str):
+        """Pre-define a collection with the given ID."""
+        assert collection_id not in [c["id"] for c in self.collections]
+        self.collections.append(
+            pystac.Collection(
+                id=collection_id,
+                description=collection_id,
+                extent=pystac.Extent(
+                    spatial=pystac.SpatialExtent([-180, -90, 180, 90]),
+                    temporal=pystac.TemporalExtent([[None, None]]),
+                ),
+                title=collection_id,
+            ).to_dict()
+        )
+
+    def predefine_item(self, *, collection_id: str, item: pystac.Item):
+        if collection_id not in [c["id"] for c in self.collections]:
+            self.predefine_collection(collection_id)
+        self.items[collection_id][item.id] = item.to_dict()
+
     def _post_collections_bulk_items(self, request, context):
         """Handler of `POST /collections/{collection_id}/bulk_items` requests."""
         # extract the collection_id from the URL
@@ -471,29 +567,31 @@ class DummyStacApi:
         post_data = request.json()
         # TODO handle insert/upsert method?
         for item_id, item in post_data["items"].items():
-            if collection_id not in self.items:
-                self.items[collection_id] = {}
             self.items[collection_id][item_id] = item
         return {}
 
     def _get_search(self, request, context):
         """Handler of `GET /search` requests."""
         collections = request.qs["collections"][0].split(",")
-        filter = request.qs["filter"][0] if "filter" in request.qs else None
-
-        if filter:
-            # TODO: use a more robust CQL2-text parser?
-            assert re.match(r"^\s*\"properties\.status\"='\w+'(\s+or\s+\"properties\.status\"='\w+')*\s*$", filter)
-            statuses = re.findall(r"\"properties\.status\"='(\w+)'", filter)
-        else:
-            statuses = None
-
         items = [
             item
             for cid in collections
             for item in self.items.get(cid, {}).values()
-            if statuses is None or item.get("properties", {}).get("status") in statuses
         ]
+        if "ids" in request.qs:
+            [ids] = request.qs["ids"]
+            ids = set(ids.split(","))
+            items = [i for i in items if i.get("id") in ids]
+        if "filter" in request.qs:
+            [property_filter] = request.qs["filter"]
+            # TODO: use a more robust CQL2-text parser?
+            assert request.qs["filter-lang"] == ["cql2-text"]
+            assert re.match(
+                r"^\s*\"properties\.status\"='\w+'(\s+or\s+\"properties\.status\"='\w+')*\s*$", property_filter
+            )
+            statuses = set(re.findall(r"\"properties\.status\"='(\w+)'", property_filter))
+            items = [i for i in items if i.get("properties", {}).get("status") in statuses]
+
         return {
             "type": "FeatureCollection",
             "features": items,
@@ -516,6 +614,38 @@ def test_run_jobs_basic(tmp_path, dummy_backend_foo, requests_mock, sleep_mock):
         }
     )
     job_db.initialize_from_df(df=df)
+    assert dummy_stac_api.items == {
+        "collection-123": {
+            "item-2024": dirty_equals.IsPartialDict(
+                {
+                    "type": "Feature",
+                    "id": "item-2024",
+                    "properties": dirty_equals.IsPartialDict(
+                        {
+                            "year": 2024,
+                            "id": None,
+                            "status": "not_started",
+                            "backend_name": None,
+                        }
+                    ),
+                }
+            ),
+            "item-2025": dirty_equals.IsPartialDict(
+                {
+                    "type": "Feature",
+                    "id": "item-2025",
+                    "properties": dirty_equals.IsPartialDict(
+                        {
+                            "year": 2025,
+                            "id": None,
+                            "status": "not_started",
+                            "backend_name": None,
+                        }
+                    ),
+                }
+            ),
+        }
+    }
 
     def create_job(row, connection, **kwargs):
         year = int(row["year"])