+++ CHANGELOG.md
@@ -9,11 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Official support for Python 3.13 (include Python 3.13 in unit test matrix on GitHub Actions) ([#653](https://github.com/Open-EO/openeo-python-client/issues/653))
+
 ### Changed
 
 ### Removed
 
 ### Fixed
+
+- `STACAPIJobDatabase.item_from()` use "datetime" from series instead of always taking "now" ([#797](https://github.com/Open-EO/openeo-python-client/issues/797))
+- Fix Python 3.13 compatibility issue in `openeo.extra.spectral_indices` ([#799](https://github.com/Open-EO/openeo-python-client/issues/799))
 
 
 ## [0.43.0] - 2025-07-02